+++ apps/dsp-app/src/app/app.module.ts
@@ -140,22 +140,7 @@ import { UriValueComponent } from './workspace/resource/values/uri-value/uri-val
 import { ListViewComponent } from './workspace/results/list-view/list-view.component';
 import { ResourceListComponent } from './workspace/results/list-view/resource-list/resource-list.component';
 import { ResultsComponent } from './workspace/results/results.component';
-import { AdvancedSearchComponent } from './workspace/search/advanced-search/advanced-search.component';
-import { ResourceAndPropertySelectionComponent } from './workspace/search/advanced-search/resource-and-property-selection/resource-and-property-selection.component';
-import { SearchSelectPropertyComponent } from './workspace/search/advanced-search/resource-and-property-selection/search-select-property/search-select-property.component';
-import { SearchBooleanValueComponent } from './workspace/search/advanced-search/resource-and-property-selection/search-select-property/specify-property-value/search-boolean-value/search-boolean-value.component';
-import { SearchDateValueComponent } from './workspace/search/advanced-search/resource-and-property-selection/search-select-property/specify-property-value/search-date-value/search-date-value.component';
-import { SearchDecimalValueComponent } from './workspace/search/advanced-search/resource-and-property-selection/search-select-property/specify-property-value/search-decimal-value/search-decimal-value.component';
-import { SearchIntValueComponent } from './workspace/search/advanced-search/resource-and-property-selection/search-select-property/specify-property-value/search-int-value/search-int-value.component';
-import { SearchLinkValueComponent } from './workspace/search/advanced-search/resource-and-property-selection/search-select-property/specify-property-value/search-link-value/search-link-value.component';
-import { SearchDisplayListComponent } from './workspace/search/advanced-search/resource-and-property-selection/search-select-property/specify-property-value/search-list-value/search-display-list/search-display-list.component';
-import { SearchListValueComponent } from './workspace/search/advanced-search/resource-and-property-selection/search-select-property/specify-property-value/search-list-value/search-list-value.component';
-import { SearchResourceComponent } from './workspace/search/advanced-search/resource-and-property-selection/search-select-property/specify-property-value/search-resource/search-resource.component';
-import { SearchTextValueComponent } from './workspace/search/advanced-search/resource-and-property-selection/search-select-property/specify-property-value/search-text-value/search-text-value.component';
-import { SearchUriValueComponent } from './workspace/search/advanced-search/resource-and-property-selection/search-select-property/specify-property-value/search-uri-value/search-uri-value.component';
-import { SpecifyPropertyValueComponent } from './workspace/search/advanced-search/resource-and-property-selection/search-select-property/specify-property-value/specify-property-value.component';
-import { SearchSelectResourceClassComponent } from './workspace/search/advanced-search/resource-and-property-selection/search-select-resource-class/search-select-resource-class.component';
-import { SearchSelectOntologyComponent } from './workspace/search/advanced-search/search-select-ontology/search-select-ontology.component';
+import { AdvancedSearchContainerComponent } from './workspace/search/advanced-search/advanced-search-container.component';
 import { ExpertSearchComponent } from './workspace/search/expert-search/expert-search.component';
 import { FulltextSearchComponent } from './workspace/search/fulltext-search/fulltext-search.component';
 import { SearchPanelComponent } from './workspace/search/search-panel/search-panel.component';
@@ -177,6 +162,7 @@ import {
 } from '@dasch-swiss/vre/shared/app-config';
 import { NgxSkeletonLoaderModule } from 'ngx-skeleton-loader';
 import { AppDatePickerModule } from '@dasch-swiss/vre/shared/app-date-picker';
+import { AdvancedSearchComponent } from '@dasch-swiss/vre/advanced-search';
 import { NgxsStoragePluginModule } from '@ngxs/storage-plugin';
 import { apiConnectionTokenProvider } from './providers/api-connection-token.provider';
 import { NgxsStoreModule } from '@dasch-swiss/vre/shared/app-state';
@@ -195,7 +181,7 @@ export function httpLoaderFactory(httpClient: HttpClient) {
         AddUserComponent,
         AddValueComponent,
         AdminImageDirective,
-        AdvancedSearchComponent,
+        AdvancedSearchContainerComponent,
         AppComponent,
         ArchiveComponent,
         AudioComponent,
@@ -265,7 +251,6 @@ export function httpLoaderFactory(httpClient: HttpClient) {
         PropertyFormComponent,
         PropertyInfoComponent,
         ReplaceFileFormComponent,
-        ResourceAndPropertySelectionComponent,
         ResourceClassFormComponent,
         ResourceClassInfoComponent,
         ResourceClassPropertyInfoComponent,
@@ -274,20 +259,7 @@ export function httpLoaderFactory(httpClient: HttpClient) {
         ResourceLinkFormComponent,
         ResourceListComponent,
         ResultsComponent,
-        SearchBooleanValueComponent,
-        SearchDateValueComponent,
-        SearchDecimalValueComponent,
-        SearchDisplayListComponent,
-        SearchIntValueComponent,
-        SearchLinkValueComponent,
-        SearchListValueComponent,
         SearchPanelComponent,
-        SearchResourceComponent,
-        SearchSelectOntologyComponent,
-        SearchSelectPropertyComponent,
-        SearchSelectResourceClassComponent,
-        SearchTextValueComponent,
-        SearchUriValueComponent,
         SelectedResourcesComponent,
         SelectGroupComponent,
         SelectLanguageComponent,
@@ -296,7 +268,6 @@ export function httpLoaderFactory(httpClient: HttpClient) {
         SelectPropertiesComponent,
         SelectResourceClassComponent,
         SortButtonComponent,
-        SpecifyPropertyValueComponent,
         SplitPipe,
         StatusComponent,
         StillImageComponent,
@@ -352,6 +323,7 @@ export function httpLoaderFactory(httpClient: HttpClient) {
         NgxSkeletonLoaderModule,
         PdfViewerModule,
         ReactiveFormsModule,
+        AdvancedSearchComponent,
         TranslateModule.forRoot({
             loader: {
                 provide: TranslateLoader,