+++ libs/vre/advanced-search/src/lib/lib.routes.ts
@@ -1,0 +1,5 @@
+import { Route } from '@angular/router';
+
+export const vreAdvancedSearchRoutes: Route[] = [
+    /* {path: '', pathMatch: 'full', component: InsertYourComponentHere} */
+];